refactor(audit): move the connection emitter into its own middleware file
`audit.middleware.ts` was at 993 lines and holds the generic `auditable()`
machinery, so the connection-creation emitter goes beside the auth and
sync-command ones instead - importing the shared actor vocabulary the same
way `auditSyncCommand.middleware.ts` does. Its tests move with it, and the
hook now imports from a connection-specific module.

`omitUndefined` is no longer needed for a metadata object whose only key is
a required string.
